YARD docs, mild refactoring
