+++ lib/neo4j.rb
@@ -59,6 +59,7 @@ require 'neo4j/active_node/persistence'
 require 'neo4j/active_node/validations'
 require 'neo4j/active_node/rels'
 require 'neo4j/active_node/reflection'
+require 'neo4j/active_node/has_n/association_cache'
 require 'neo4j/active_node/has_n'
 require 'neo4j/active_node/has_n/association'
 require 'neo4j/active_node/query/query_proxy'
