+++ lib/neo4j/active_node/dependent.rb
@@ -1,11 +1,13 @@
 module Neo4j
   module ActiveNode
     module Dependent
+      # When a `dependent: :destroy` action occurs, this array holds objects preparing for deletion.
       def dependent_children
         @dependent_children ||= []
       end
 
-      attr_writer :called_by
+      # \@called_by is used association dependency to prevent loops
+      attr_accessor :called_by
     end
   end
 end
