+++ lib/neo4j/active_node/dependent/query_proxy_methods.rb
@@ -5,7 +5,7 @@ module Neo4j
       module QueryProxyMethods
         # Used as part of `dependent: :destroy` and may not have any utility otherwise.
         # It keeps track of the node responsible for a cascading `destroy` process.
-        # @param [#dependent_children] caller The node that called this method. Typically, we would use QueryProxy's `caller` method
+        # @param [#dependent_children] owning_node The node that called this method. Typically, we would use QueryProxy's `caller` method
         # but this is not always available, so we require it explicitly.
         def each_for_destruction(owning_node)
           target = owning_node.called_by || owning_node
