+++ lib/neo4j/type_converters.rb
@@ -1,6 +1,6 @@
 module Neo4j
   module TypeConverters
-    # This exists for legacy purposes. Some gems (*coughDevisecough*) that the Neo4jrb project does not own
+    # This exists for legacy purposes. Some gems that the Neo4jrb project does not own
     # may contain references to this file. We will remove it once that has been dealt with.
     include Neo4j::Shared::TypeConverters
   end
